+++ sdmx/__init__.py
@@ -1,6 +1,5 @@
 import logging
-
-import pkg_resources
+from importlib.metadata import PackageNotFoundError, version
 
 from sdmx.client import Client, Request, read_url
 from sdmx.reader import read_sdmx
@@ -23,8 +22,8 @@ __all__ = [
 
 
 try:
-    __version__ = pkg_resources.get_distribution("sdmx1").version
-except Exception:  # pragma: no cover
+    __version__ = version("sdmx1")
+except PackageNotFoundError:  # pragma: no cover
     # Local copy or not installed with setuptools
     __version__ = "999"